+++ src/main/java/com/kuit/healthmate/domain/supplement/SupplementChecker.java
@@ -1,4 +1,4 @@
-package com.kuit.healthmate.domain;
+package com.kuit.healthmate.domain.supplement;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,21 +8,22 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.time.LocalTime;
+import java.time.LocalDate;
 
 @Entity
 @Table
-public class SupplementTime {
+public class SupplementChecker {
+
     @Id
     @GeneratedValue
-    @Column(name = "supplement_time_id")
+    @Column(name = "supplement_checker_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="supplement_id")
     private Supplement supplement;
 
+    private LocalDate checkDate;
     private TimeSlot timeSlot;
-
-    private LocalTime time;
+    private Boolean status;
 }
