+++ src/main/java/com/kuit/healthmate/domain/supplement/SupplementStatus.java
@@ -1,4 +1,4 @@
-package com.kuit.healthmate.domain;
+package com.kuit.healthmate.domain.supplement;
 
 public enum SupplementStatus {
     ACTIVE, DORMANT, DELETED
