+++ src/main/java/com/kuit/healthmate/domain/supplement/SupplementTime.java
@@ -1,4 +1,4 @@
-package com.kuit.healthmate.domain;
+package com.kuit.healthmate.domain.supplement;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,22 +8,21 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.time.LocalDate;
+import java.time.LocalTime;
 
 @Entity
 @Table
-public class SupplementChecker {
-
+public class SupplementTime {
     @Id
     @GeneratedValue
-    @Column(name = "supplement_checker_id")
+    @Column(name = "supplement_time_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="supplement_id")
     private Supplement supplement;
 
-    private LocalDate checkDate;
     private TimeSlot timeSlot;
-    private Boolean status;
+
+    private LocalTime time;
 }
